Restructure the `package.py` script into a python module.
