+++ olca-app-build/package/__main__.py
@@ -1,0 +1,7 @@
+import sys
+
+
+if __name__ == "__main__":
+    from package.main import main as _main
+
+    sys.exit(_main())
